extract serial to context
